refactor: split components
