+++ src/features/form/index.module.scss
@@ -1,0 +1,126 @@
+@use './app/styles/mixins.scss' as *;
+
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 40rem;
+  margin: 0 auto;
+  padding: 12rem 0;
+}
+
+.input {
+  @include poppins;
+  padding: 0.8rem;
+  font-size: 1.6rem;
+  border-radius: 0.6rem;
+  border: 2px solid var(--dark-l);
+  margin: 1rem 0 2.08rem;
+  background-color: var(--light);
+
+  &:focus-visible {
+    outline: none;
+    border: 2px solid var(--dark-xxl);
+  }
+}
+
+.inputError {
+  border: 2px solid var(--error);
+  margin-bottom: 0.5rem;
+}
+
+.label {
+  margin-top: 1.2rem;
+}
+
+.error {
+  color: var(--error);
+}
+
+.agreement {
+  line-height: 1;
+  font-size: 1.4rem;
+}
+
+.checkbox {
+  margin-right: 1rem;
+}
+
+.button {
+  @include poppins($color: var(--dark-xl));
+  margin-top: 2rem;
+  padding: 1.5rem;
+  border-radius: 0.5rem;
+  background-color: var(--light);
+  border: 0.2rem solid var(--dark-xl);
+  cursor: pointer;
+  transition: all 0.3s ease-in-out;
+
+  &:hover {
+    color: var(--dark-xxl);
+    background-color: var(--primary);
+  }
+
+  &:disabled {
+    cursor: auto;
+    background-color: var(--dark-s);
+    border: 0.2rem solid var(--dark-l);
+
+    &:hover {
+      $color: var(--dark-xl);
+      background-color: var(--dark-s);
+    }
+  }
+}
+
+// switcher
+
+.switcherWrapper {
+  width: 100%;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  margin-top: 1rem;
+}
+
+.switcher {
+  height: 0;
+  width: 0;
+  visibility: hidden;
+}
+
+.switcherLabel {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  cursor: pointer;
+  width: 5rem;
+  height: 2.6rem;
+  background: var(--light);
+  border-radius: 10rem;
+  border: 2px solid var(--dark-l);
+  position: relative;
+  transition: background-color 0.2s;
+  margin-right: 1rem;
+}
+
+.switcherLabel .switcherBtn {
+  content: '';
+  position: absolute;
+  top: 0.3rem;
+  left: 0.2rem;
+  width: 2rem;
+  height: 2rem;
+  border-radius: 50%;
+  transition: 0.2s;
+  background: var(--dark-l);
+}
+
+.switcher:checked + .switcherLabel .switcherBtn {
+  left: calc(100% - 2px);
+  transform: translateX(-100%);
+}
+
+.switcher:checked + .switcherLabel {
+  background: var(--primary);
+}
