+++ src/features/form/models.ts
@@ -1,0 +1,8 @@
+export enum UserInput {
+  LAST_NAME = 'lastName',
+  FIRST_NAME = 'firstName',
+  BIRTHDAY = 'birthday',
+  COUNTRY = 'country',
+  AVATAR = 'avatar',
+  AGREEMENT = 'agreement',
+}
