+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import { Route, Routes, Navigate } from 'react-router-dom';
 import About from 'pages/about';
 import Home from 'pages/home';
 import NotFound from 'pages/404';
-import Form from 'pages/form';
+import FormPage from 'pages/form';
 
 class Routing extends React.Component {
   render() {
@@ -12,7 +12,7 @@ class Routing extends React.Component {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
-        <Route path="/form" element={<Form />} />
+        <Route path="/form" element={<FormPage />} />
         <Route path="/404" element={<NotFound />} />
         <Route path="*" element={<Navigate to="/404" />} />
       </Routes>
